Gym 1 models

See merge request graduation_project_5th/graduation-project!5

+++ database/migrations/2023_12_27_182606_create_subscribes_table.php
@@ -14,7 +14,7 @@ class CreateSubscribesTable extends Migration
             $table->bigInteger('plan_id')->unsigned();
             $table->bigInteger('coach_Id')->unsigned();
             $table->bigInteger('offer_id')->unsigned();
-            $table->double('cost');
+            $table->decimal('cost', 10, 2);
             $table->text('description');
             $table->date('start_date');
             $table->date('end_date');
